Phone-system-api-create-user-from-admin-and-checkout-page

+++ database/migrations/2020_10_15_123403_alter_add_customer_id_dashboard_link_phone_system_fancy_numbers.php
@@ -14,7 +14,7 @@ class AlterAddCustomerIdDashboardLinkPhoneSystemFancyNumbers extends Migration
     public function up()
     {
         Schema::table('fancy_numbers', function (Blueprint $table) {
-            $table->unsignedInteger('customer_id_phone_system')->nullable();
+            $table->unsignedBigInteger('customer_id_phone_system')->nullable();
             $table->string('dashboard_link_phone_system')->nullable();
         });
     }
